my-app/requirements.txt
flask==2.3.3
flask-cors==4.0.0
ultralytics==8.0.196
pillow==10.0.1
numpy==1.24.3
